Hide the commit hash found in metric descriptions (#20)
The azure/online-experimentation-deploy-metrics action has an `add-commit-hash-to-metric-description` input that appended the git commit hash to the description of each deployed metric. This is a useful feature for metric versioning, but it's confusing when reviewing the analysis results in the markdown summary. We now strip this info before showing the metric description in the markdown.